MAX7219_Digital_Clock/requirements.txt
luma.led_matrix
paho-mqtt
spidev
flask
requests
feedparser
pillow